Fixed the way data is handled when passed to the next view
The data is passed to the Status view indirectly by using stateStorage.

We're using two variables, one is the repo URL and the other one is a
bool that indicates if state is being used or not. If the state is used,
we will need to send it in the request while getting status. Else we
will just send the request.